refactor(admin): fold connection manager view and logic into one module
The IMS connection manager page and its session-connection logic were split between view/, composables/ and styles/. Move ConnectionManager.vue, useConnectionManager.ts and its scss under admin/connection-manager to keep the module self-contained.
