+++ webui/src/admin/connection-manager/connection-manager.scss
